DSL-server/requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
packaging==24.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.15
pywebio==1.8.3
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
SQLAlchemy==2.0.36
tornado==6.4.2
typing_extensions==4.12.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
Werkzeug==3.1.3
wheel==0.44.0
